# api-docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Time Here Now API",
    "version": "20251023",
    "description": "API to get the current time based on timezone or client IP. **IMPORTANT: All time values are sourced from NEAR blockchain, NOT system/NTP time.** The API polls NEAR blockchain at 5 Hz (200ms intervals) and serves cached blockchain timestamps. Time zones and offsets are sourced from the IANA Time Zone Database (tzdb). Returns HTTP 503 if blockchain time is unavailable."
  },
  "servers": [
    { "url": "https://timeherenow.rodit.org:8443/api" },
    { "url": "http://localhost:8080/api" }
  ],
...
  "paths": {
...
    "/mcp/resources": {
      "get": {
        "summary": "List available MCP resources",
        "description": "Returns a list of resources available through the Model Context Protocol interface",
        "tags": ["MCP"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer" },
            "description": "Maximum number of resources to return"
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": { "type": "string" },
            "description": "Cursor for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "List of available resources with pagination info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resources": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "uri": { "type": "string" },
                          "name": { "type": "string" },
                          "type": { "type": "string" }
                        }
                      }
                    },
                    "nextCursor": { "type": "string", "nullable": true },
                    "requestId": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorJsonResponse" },
          "500": { "$ref": "#/components/responses/ErrorJsonResponse" }
        }
      }
    },
    "/mcp/resource/{uri}": {
      "get": {
        "summary": "Get a specific MCP resource",
        "description": "Returns the content of a specific resource",
        "tags": ["MCP"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "uri",
            "required": true,
            "schema": { "type": "string" },
            "description": "URI of the resource to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Resource content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": { "type": "string" },
                    "content": { "type": "object" },
                    "requestId": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorJsonResponse" },
          "404": { "$ref": "#/components/responses/ErrorJsonResponse" },
          "500": { "$ref": "#/components/responses/ErrorJsonResponse" }
        }
      }
    },
    "/mcp/schema": {
      "get": {
        "summary": "Get MCP schema",
        "description": "Returns the OpenAPI schema for the MCP interface",
        "tags": ["MCP"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "MCP schema",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorJsonResponse" },
          "500": { "$ref": "#/components/responses/ErrorJsonResponse" }
        }
      }
    },
...
  },
  "components": {
    "responses": {
...
      "ErrorJsonResponse": {
        "description": "an error response in JSON format",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorJsonResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorJsonResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "details about the error encountered"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}